Move link component to elementLink
