+++ assets/src/edit-story/components/canvas/frameElement.js
@@ -33,7 +33,7 @@ import {
 } from '../../elements/shared';
 import { useUnits } from '../../units';
 import WithMask from '../../masks/frame';
-import WithLink from '../link/frame';
+import WithLink from '../elementLink/frame';
 import { useTransformHandler } from '../transform';
 import useCanvas from './useCanvas';
