+++ assets/src/edit-story/components/panels/link/index.js
@@ -31,7 +31,7 @@ import { __ } from '@wordpress/i18n';
  */
 import { useDebouncedCallback } from 'use-debounce';
 import { Media, Row, Button } from '../../form';
-import { createLink, getLinkFromElement } from '../../link';
+import { createLink, getLinkFromElement } from '../../elementLink';
 import { useAPI } from '../../../app/api';
 import { useSnackbar } from '../../../app/snackbar';
 import { isValidUrl, toAbsoluteUrl, withProtocol } from '../../../utils/url';
